// src/types/flashcards.ts
export const NATIVE_LANGUAGES = [
  { value: 'Arabic', label: 'Arabic', flag: '🇸🇦' },
  { value: 'Bengali', label: 'Bengali', flag: '🇧🇩' },
  { value: 'Bulgarian', label: 'Bulgarian', flag: '🇧🇬' },
  { value: 'Catalan', label: 'Catalan', flag: '🇪🇸' },
  { value: 'Chinese', label: 'Chinese', flag: '🇨🇳' },
  { value: 'Croatian', label: 'Croatian', flag: '🇭🇷' },
  { value: 'Czech', label: 'Czech', flag: '🇨🇿' },
  { value: 'Danish', label: 'Danish', flag: '🇩🇰' },
  { value: 'Dutch', label: 'Dutch', flag: '🇳🇱' },
  { value: 'Finnish', label: 'Finnish', flag: '🇫🇮' },
  { value: 'French', label: 'French', flag: '🇫🇷' },
  { value: 'German', label: 'German', flag: '🇩🇪' },
  { value: 'Greek', label: 'Greek', flag: '🇬🇷' },
  { value: 'Hebrew', label: 'Hebrew', flag: '🇮🇱' },
  { value: 'Hindi', label: 'Hindi', flag: '🇮🇳' },
  { value: 'Hungarian', label: 'Hungarian', flag: '🇭🇺' },
  { value: 'Indonesian', label: 'Indonesian', flag: '🇮🇩' },
  { value: 'Italian', label: 'Italian', flag: '🇮🇹' },
  { value: 'Japanese', label: 'Japanese', flag: '🇯🇵' },
  { value: 'Javanese', label: 'Javanese', flag: '🇮🇩' },
  { value: 'Korean', label: 'Korean', flag: '🇰🇷' },
  { value: 'Malay', label: 'Malay', flag: '🇲🇾' },
  { value: 'Marathi', label: 'Marathi', flag: '🇮🇳' },
  { value: 'Norwegian', label: 'Norwegian', flag: '🇳🇴' },
  { value: 'Persian', label: 'Persian', flag: '🇮🇷' },
  { value: 'Polish', label: 'Polish', flag: '🇵🇱' },
  { value: 'Portuguese', label: 'Portuguese', flag: '🇵🇹' },
  { value: 'Punjabi', label: 'Punjabi', flag: '🇮🇳' },
  { value: 'Romanian', label: 'Romanian', flag: '🇷🇴' },
  { value: 'Russian', label: 'Russian', flag: '🇷🇺' },
  { value: 'Serbian', label: 'Serbian', flag: '🇷🇸' },
  { value: 'Slovak', label: 'Slovak', flag: '🇸🇰' },
  { value: 'Slovenian', label: 'Slovenian', flag: '🇸🇮' },
  { value: 'Spanish', label: 'Spanish', flag: '🇪🇸' },
  { value: 'Swahili', label: 'Swahili', flag: '🇰🇪' },
  { value: 'Swedish', label: 'Swedish', flag: '🇸🇪' },
  { value: 'Tagalog', label: 'Tagalog', flag: '🇵🇭' },
  { value: 'Tamil', label: 'Tamil', flag: '🇮🇳' },
  { value: 'Telugu', label: 'Telugu', flag: '🇮🇳' },
  { value: 'Thai', label: 'Thai', flag: '🇹🇭' },
  { value: 'Turkish', label: 'Turkish', flag: '🇹🇷' },
  { value: 'Ukrainian', label: 'Ukrainian', flag: '🇺🇦' },
  { value: 'Urdu', label: 'Urdu', flag: '🇵🇰' },
  { value: 'Vietnamese', label: 'Vietnamese', flag: '🇻🇳' },
] as const;